+++ network/validator/validator.go
@@ -1,0 +1,16 @@
+package validator
+
+import (
+	"context"
+
+	"github.com/libp2p/go-libp2p-core/peer"
+	pubsub "github.com/libp2p/go-libp2p-pubsub"
+)
+
+// MessageValidator validates the given message with original sender `from` and returns an error if validation fails.
+// Note: contrarily to pubsub.ValidatorEx, the peerID parameter does not represent the bearer of the message, but its source.
+type MessageValidator func(ctx context.Context, from peer.ID, msg interface{}) error
+
+// PubSubMessageValidator validates the given message with original sender `from` and returns pubsub.ValidationResult.
+// Note: contrarily to pubsub.ValidatorEx, the peerID parameter does not represent the bearer of the message, but its source.
+type PubSubMessageValidator func(ctx context.Context, from peer.ID, msg interface{}) pubsub.ValidationResult
